Refactor dwio and common io modules [1/n]
dwio modules and common io modules are interwined.
This is 1/n PR attempting to refactor common-io libararies out from dwio.

Current module design in:

In this PR,
- CachedBufferedInput has been moved to velox/common/io folder, without the
change to namespace.
- CachedBufferedInput has been de-coupled from MetricsLog, MetricsType and
ReaderOptions.
