+++ velox/common/io/CMakeLists.txt
@@ -11,16 +11,9 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-add_subdirectory(base)
-add_subdirectory(caching)
-add_subdirectory(compression)
-add_subdirectory(config)
-add_subdirectory(encode)
-add_subdirectory(file)
-add_subdirectory(hyperloglog)
-add_subdirectory(io)
-add_subdirectory(memory)
-add_subdirectory(process)
-add_subdirectory(serialization)
-add_subdirectory(time)
-add_subdirectory(testutil)
+add_library(velox_common_io CachedBufferedInput.cpp CacheInputStream.cpp)
+
+target_include_directories(velox_common_io PRIVATE ${Protobuf_INCLUDE_DIRS})
+
+target_link_libraries(velox_common_io xsimd velox_dwio_common Folly::folly
+                      glog::glog)
