+++ velox/dwio/dwrf/test/CacheInputTest.cpp
@@ -20,6 +20,7 @@
 #include "velox/common/caching/FileIds.h"
 #include "velox/common/file/FileSystems.h"
 #include "velox/common/memory/MmapAllocator.h"
+#include "velox/dwio/common/CacheInputStream.h"
 #include "velox/dwio/common/CachedBufferedInput.h"
 #include "velox/dwio/dwrf/common/Common.h"
 #include "velox/exec/tests/utils/TempDirectoryPath.h"
